feat: upload and delete post images, aws S3 integration, bucket configuration

+++ services/tags.service.js
@@ -98,7 +98,8 @@ class TagsServices {
       let tag = await models.Tags.findByPk(id)
 
       if (!tag) throw new CustomError('Not found Tag', 404, 'Not Found')
-
+      if (tag.image_url) throw new CustomError('Image Tag is on Cloud, must be deleted first', 400, 'Bad Request')
+      
       await tag.destroy({ transaction })
 
       await transaction.commit()
